[circuit-writer] factor out prover and constants from writer mod
