+++ circuit-construction/src/writer.rs
@@ -1,108 +1,13 @@
-use ark_ec::{AffineCurve, ProjectiveCurve};
-use ark_ff::{BigInteger, FftField, Field, One, PrimeField, SquareRootField, Zero};
+use ark_ff::{BigInteger, FftField, PrimeField};
 use array_init::array_init;
-use commitment_dlog::{
-    commitment::{CommitmentCurve, PolyComm},
-    srs::{endos, SRS},
+use kimchi::circuits::{
+    gate::{CircuitGate, GateType},
+    wires::{Wire, COLUMNS},
 };
-use kimchi::{
-    circuits::{
-        constraints::ConstraintSystem,
-        gate::{CircuitGate, GateType},
-        wires::{Wire, COLUMNS},
-    },
-    plonk_sponge::FrSponge,
-    proof::ProverProof,
-    prover_index::ProverIndex,
-};
-use mina_curves::pasta::{fp::Fp, fq::Fq, pallas::Affine as Other, vesta::Affine};
-use oracle::{constants::*, permutation::full_round, poseidon::ArithmeticSpongeParams, FqSponge};
+use oracle::{constants::*, permutation::full_round};
 use std::collections::HashMap;
 
-pub const GENERICS: usize = 3;
-
-pub const SINGLE_GENERIC_COEFFS: usize = 5;
-pub const GENERIC_ROW_COEFFS: usize = 2 * SINGLE_GENERIC_COEFFS;
-
-pub trait Cycle {
-    type InnerField: FftField
-        + PrimeField
-        + SquareRootField
-        + From<u128>
-        + From<u64>
-        + From<u32>
-        + From<u16>
-        + From<u8>;
-    type OuterField: FftField
-        + PrimeField
-        + SquareRootField
-        + From<u128>
-        + From<u64>
-        + From<u32>
-        + From<u16>
-        + From<u8>;
-
-    type InnerMap: groupmap::GroupMap<Self::InnerField>;
-    type OuterMap: groupmap::GroupMap<Self::OuterField>;
-
-    type InnerProj: ProjectiveCurve<
-            Affine = Self::Inner,
-            ScalarField = Self::OuterField,
-            BaseField = Self::InnerField,
-        > + From<Self::Inner>
-        + Into<Self::Inner>
-        + std::ops::MulAssign<Self::OuterField>;
-
-    type Inner: CommitmentCurve<
-            Projective = Self::InnerProj,
-            Map = Self::InnerMap,
-            BaseField = Self::InnerField,
-            ScalarField = Self::OuterField,
-        > + From<Self::InnerProj>
-        + Into<Self::InnerProj>;
-
-    type OuterProj: ProjectiveCurve<
-            Affine = Self::Outer,
-            ScalarField = Self::InnerField,
-            BaseField = Self::OuterField,
-        > + From<Self::Outer>
-        + Into<Self::Outer>
-        + std::ops::MulAssign<Self::InnerField>;
-
-    type Outer: CommitmentCurve<
-        Projective = Self::OuterProj,
-        Map = Self::OuterMap,
-        ScalarField = Self::InnerField,
-        BaseField = Self::OuterField,
-    >;
-}
-
-pub struct FpInner;
-pub struct FqInner;
-
-impl Cycle for FpInner {
-    type InnerMap = <Other as CommitmentCurve>::Map;
-    type OuterMap = <Affine as CommitmentCurve>::Map;
-
-    type InnerField = Fp;
-    type OuterField = Fq;
-    type Inner = Other;
-    type Outer = Affine;
-    type InnerProj = <Other as AffineCurve>::Projective;
-    type OuterProj = <Affine as AffineCurve>::Projective;
-}
-
-impl Cycle for FqInner {
-    type InnerMap = <Affine as CommitmentCurve>::Map;
-    type OuterMap = <Other as CommitmentCurve>::Map;
-
-    type InnerField = Fq;
-    type OuterField = Fp;
-    type Inner = Affine;
-    type Outer = Other;
-    type InnerProj = <Affine as AffineCurve>::Projective;
-    type OuterProj = <Other as AffineCurve>::Projective;
-}
+use crate::constants::{Constants, GENERICS, GENERIC_ROW_COEFFS};
 
 #[derive(Hash, Eq, PartialEq, Debug, Clone, Copy)]
 pub struct Var<F> {
@@ -122,13 +27,6 @@ pub struct GateSpec<F: FftField> {
     pub typ: GateType,
     pub row: [Var<F>; COLUMNS],
     pub c: Vec<F>,
-}
-
-#[derive(Clone)]
-pub struct Constants<F: Field> {
-    pub poseidon: ArithmeticSpongeParams<F>,
-    pub endo: F,
-    pub base: (F, F),
 }
 
 pub struct System<F: FftField> {
@@ -793,7 +691,7 @@ impl<F: FftField + PrimeField> Cs<F> for WitnessGenerator<F> {
 
 impl<F: FftField> WitnessGenerator<F> {
     /// Returns the columns of the witness.
-    fn columns(&self) -> [Vec<F>; COLUMNS] {
+    pub fn columns(&self) -> [Vec<F>; COLUMNS] {
         array_init(|col| self.rows.iter().map(|row| row[col]).collect())
     }
 }
@@ -859,154 +757,5 @@ impl<F: FftField> System<F> {
         }
 
         gates
-    }
-}
-
-/// Given an index, a group map, custom blinders for the witness, a public input vector, and a circuit `main`, create a proof.
-pub fn prove<G, H, EFqSponge, EFrSponge>(
-    index: &ProverIndex<G>,
-    group_map: &G::Map,
-    blinders: Option<[Option<G::ScalarField>; COLUMNS]>,
-    public_input: Vec<G::ScalarField>,
-    mut main: H,
-) -> ProverProof<G>
-where
-    H: FnMut(&mut WitnessGenerator<G::ScalarField>, Vec<Var<G::ScalarField>>),
-    G::BaseField: PrimeField,
-    G: CommitmentCurve,
-    EFqSponge: Clone + FqSponge<G::BaseField, G, G::ScalarField>,
-    EFrSponge: FrSponge<G::ScalarField>,
-{
-    // create the public rows
-    let mut gen: WitnessGenerator<G::ScalarField> = WitnessGenerator {
-        rows: public_input
-            .iter()
-            .map(|x| array_init(|i| if i == 0 { *x } else { G::ScalarField::zero() }))
-            .collect(),
-    };
-
-    // run the witness generation
-    let public_vars = public_input
-        .iter()
-        .map(|x| Var {
-            index: 0,
-            value: Some(*x),
-        })
-        .collect();
-    main(&mut gen, public_vars);
-
-    // get the witness columns
-    let columns = gen.columns();
-
-    // custom blinders for the witness commitment
-    let blinders: [Option<PolyComm<G::ScalarField>>; COLUMNS] = match blinders {
-        None => array_init(|_| None),
-        Some(bs) => array_init(|i| {
-            bs[i].map(|b| PolyComm {
-                unshifted: vec![b],
-                shifted: None,
-            })
-        }),
-    };
-
-    // create the proof
-    ProverProof::create_recursive::<EFqSponge, EFrSponge>(
-        group_map,
-        columns,
-        &[],
-        index,
-        vec![],
-        Some(blinders),
-    )
-    .unwrap()
-}
-
-pub fn generate_prover_index<C, H>(
-    srs: std::sync::Arc<SRS<C::Outer>>,
-    constants: &Constants<C::InnerField>,
-    poseidon_params: &ArithmeticSpongeParams<C::OuterField>,
-    public: usize,
-    main: H,
-) -> ProverIndex<C::Outer>
-where
-    H: FnOnce(&mut System<C::InnerField>, Vec<Var<C::InnerField>>),
-    C: Cycle,
-{
-    let mut system: System<C::InnerField> = System {
-        next_variable: 0,
-        gates: vec![],
-    };
-    let z = C::InnerField::zero();
-
-    // create public input variables
-    let public_input_row = vec![C::InnerField::one(), z, z, z, z, z, z, z, z, z];
-    let public_input: Vec<_> = (0..public)
-        .map(|_| {
-            let v = system.var(|| panic!("fail"));
-            let row = array_init(|i| {
-                if i == 0 {
-                    v
-                } else {
-                    system.var(|| panic!("fail"))
-                }
-            });
-            system.gate(GateSpec {
-                typ: GateType::Generic,
-                c: public_input_row.clone(),
-                row,
-            });
-            v
-        })
-        .collect();
-
-    main(&mut system, public_input);
-
-    let gates = system.gates();
-    println!("gates: {}", gates.len());
-    // Other base field = self scalar field
-    let (endo_q, _endo_r) = endos::<C::Inner>();
-
-    let constraint_system =
-        ConstraintSystem::<C::InnerField>::create(gates, constants.poseidon.clone())
-            .public(public)
-            .build()
-            // TODO: return a Result instead of panicking
-            .expect("couldn't construct constraint system");
-
-    ProverIndex::<C::Outer>::create(constraint_system, poseidon_params.clone(), endo_q, srs)
-}
-
-pub fn fp_constants() -> Constants<Fp> {
-    let (endo_q, _endo_r) = endos::<Other>();
-    let base = Other::prime_subgroup_generator().to_coordinates().unwrap();
-    Constants {
-        poseidon: oracle::pasta::fp_kimchi::params(),
-        endo: endo_q,
-        base,
-    }
-}
-
-pub fn fq_constants() -> Constants<Fq> {
-    let (endo_q, _endo_r) = endos::<Affine>();
-    let base = Affine::prime_subgroup_generator().to_coordinates().unwrap();
-    Constants {
-        poseidon: oracle::pasta::fq_kimchi::params(),
-        endo: endo_q,
-        base,
-    }
-}
-
-pub fn shift<F: PrimeField>(size: usize) -> F {
-    let two: F = 2_u64.into();
-    two.pow(&[size as u64])
-}
-
-pub trait CoordinateCurve: AffineCurve {
-    fn to_coords(&self) -> Option<(Self::BaseField, Self::BaseField)>;
-}
-
-impl<G: CommitmentCurve> CoordinateCurve for G {
-    fn to_coords(&self) -> Option<(Self::BaseField, Self::BaseField)> {
-        CommitmentCurve::to_coordinates(self)
     }
 }
